change all int to long

+++ project.cpp
@@ -8,15 +8,15 @@ using namespace std;
 
 typedef struct vertice
 {
-    int parent;
-    int rank;
+    long parent;
+    long rank;
 } Vertice;
 
 typedef struct edge
 {
-    int vertice1;
-    int vertice2;
-    int weight;
+    long vertice1;
+    long vertice2;
+    long weight;
 } Edge;
 
 /* Global vectors */
@@ -26,20 +26,20 @@ vector<Edge> edges;
 
 /* Functions*/
 
-void makeSet(int V)
+void makeSet(long V)
 {
     vertices[V].parent = V;
     vertices[V].rank = 0;
 }
 
-int findSet(int V)
+long findSet(long V)
 {
     if (vertices[V].parent != V)
         vertices[V].parent = findSet(vertices[V].parent);
     return vertices[V].parent;
 }
 
-void link(int v1, int v2)
+void link(long v1, long v2)
 {
     if (vertices[v1].rank > vertices[v2].rank)
         vertices[v2].parent = v1;
@@ -51,7 +51,7 @@ void link(int v1, int v2)
     }
 }
 
-void treeUnion(int v1, int v2)
+void treeUnion(long v1, long v2)
 {
     link(findSet(v1), findSet(v2));
 }
@@ -63,9 +63,9 @@ bool weightComparator(Edge e1, Edge e2)
 
 // Performs Kruskal's algorithm but from heavist to lightest, and instead of
 // adding edges to the MST, it simply adds their weight to "sum"
-int adaptedKruskal()
+long adaptedKruskal()
 {
-    int sum = 0;
+    long sum = 0;
 
     // Sorts all edges from heaviest to lightest
     std::sort(edges.begin(), edges.end(), weightComparator);
@@ -80,9 +80,9 @@ int adaptedKruskal()
     return sum;
 }
 
-int main()
+long main()
 {
-    int numVertices, numEdges, v1, v2, weight;
+    long numVertices, numEdges, v1, v2, weight;
     if (scanf("%d %d", &numVertices, &numEdges) != 2)
     {
         exit(1);
@@ -93,10 +93,10 @@ int main()
 
     // Performs "makeSet" on every vertice.
     // The vertice "1" is on vertices[0], "2" is on vertices[1], etc.
-    for (int i = 0; i < numVertices; i++)
+    for (long i = 0; i < numVertices; i++)
         makeSet(i);
 
-    for (int i = 0; i < numEdges; i++)
+    for (long i = 0; i < numEdges; i++)
     {
         if (scanf("%d %d %d", &v1, &v2, &weight) != 3)
         {
